Updated stereoRig.getFundamentalMatrix() using formula from Multiple View Geometry in Computer Vision, by Richard Hartley and Andrew Zisserman

+++ simplestereo/utils.py
@@ -216,7 +216,7 @@ def getCrossProductMatrix(v):
     """
     return np.array( [ [0, -v[2], v[1]], \
                        [v[2], 0, -v[0]], \
-                       [-v[1], v[0], 0] ] )
+                       [-v[1], v[0], 0] ] , dtype=np.float)
